(continuacao): Vaklidacoes de Formulario no Front-end

+++ src/app/courses/containers/courses/courses.component.ts
@@ -3,67 +3,72 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { catchError, Observable, of } from 'rxjs';
+import { ConfimationDialogComponent } from 'src/app/shared/components/confimation-dialog/confimation-dialog.component';
 import { ErroDialogComponent } from 'src/app/shared/components/erro-dialog/erro-dialog.component';
 
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
 
-
 @Component({
   selector: 'app-courses',
   templateUrl: './courses.component.html',
-  styleUrls: ['./courses.component.scss']
+  styleUrls: ['./courses.component.scss'],
 })
-
 export class CoursesComponent {
-
-  courses$:Observable<Course[]> | null = null;
+  courses$: Observable<Course[]> | null = null;
 
   constructor(
-    private coursesService:CoursesService,
+    private coursesService: CoursesService,
     public dialog: MatDialog,
     private router: Router,
     private snackBar: MatSnackBar,
-    private route:ActivatedRoute
-    ) {
+    private route: ActivatedRoute
+  ) {
     this.refresh();
   }
 
-  refresh(){
-      this.courses$ = this.coursesService.list()
-      .pipe(
-        catchError(error => {
-          this.onError("Erro ao carregar cursos")
-          return of([])
-        })
-      );
+  refresh() {
+    this.courses$ = this.coursesService.list().pipe(
+      catchError((error) => {
+        this.onError('Erro ao carregar cursos');
+        return of([]);
+      })
+    );
   }
 
-  onError(errorMsg:string) {
+  onError(errorMsg: string) {
     this.dialog.open(ErroDialogComponent, {
-      data: errorMsg
+      data: errorMsg,
     });
   }
 
-  onAdd(){
-    this.router.navigate(['new'], {relativeTo:this.route});
+  onAdd() {
+    this.router.navigate(['new'], { relativeTo: this.route });
   }
 
   onEdit(course: Course) {
-    this.router.navigate(['edit', course._id],{relativeTo:this.route});
+    this.router.navigate(['edit', course._id], { relativeTo: this.route });
   }
 
-  onRemove(course:Course) {
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open("Curso removido com sucesso!",'X',{
-          duration:5000,
-          verticalPosition: "top",
-          horizontalPosition: "center"
-        });
-      },
-      () => this.onError("Erro ao tentar remover curso!")
-    );
+  onRemove(course: Course) {
+    const dialogRef = this.dialog.open(ConfimationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Curso removido com sucesso!', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
+          },
+          () => this.onError('Erro ao tentar remover curso!')
+        );
+      }
+    });
   }
 }
